Add dynamic schema on translation tests
